+++ src/common/enums.ts
@@ -6,6 +6,11 @@ export const LanguageName: Record<Language, string> = {
     pl: 'Polish'
 };
 
+export enum DishProvider {
+    INT_DMT_USER = 'int_dmt_user',                  // NOTE: DishMatcher user
+    EXT_API_SPOONACULAR = 'ext_api_spoonacular',    // NOTE: External API - Spoonacular
+}
+
 export enum MealType {
     ANY = 'any',
     BREAKFAST = 'breakfast',