+++ src/common/guards.ts
@@ -1,0 +1,5 @@
+import { DishRepository } from '../mongodb/repositories/dish.repository';
+
+export function isDishRepository(dishRepository: unknown): dishRepository is DishRepository {
+    return dishRepository instanceof DishRepository;
+}