+++ src/services/abstract.api.service.ts
@@ -1,18 +1,20 @@
 import { Injectable } from '@nestjs/common';
 import { RedisService } from '../modules/redis/redis.service';
-import { MealType } from '../common/enums';
+import { DishProvider, MealType } from '../common/enums';
 import { DetailedDish, DishRecipeSections, RatedDish } from '../modules/dish/dish.types';
 import { getQueryWithIngredientsAndDishType } from '../modules/dish/dish.utils';
 import { AxiosResponse } from 'axios';
 import { ApiName } from '../modules/redis/redis.types';
-import { ContextString, Language } from '../common/types';
+import { ContextString, EncodedDishId, Language } from '../common/types';
 import { LoggerService } from '../modules/logger/logger.service';
 import { AxiosService } from './axios.service';
 import { IngredientType, DishIngredient } from '../modules/ingredient/ingredient.types';
 import { DishRecipe } from '../modules/recipe/recipe.types';
+import { DishProvidable } from '../common/interfaces';
+import { DishIdObfuscator } from '../common/helpers/dish-id-obfuscator.helper';
 
 @Injectable()
-export abstract class AbstractApiService<GenericDishStruct, GenericIngredientStruct, GenericDishDetailsStruct, DishInstructionStruct> {
+export abstract class AbstractApiService<GenericDishStruct, GenericIngredientStruct, GenericDishDetailsStruct, DishInstructionStruct> implements DishProvidable {
 
     constructor(
         protected readonly axiosService: AxiosService,
@@ -30,6 +32,7 @@ export abstract class AbstractApiService<GenericDishStruct, GenericIngredientStr
 
     abstract getDishInstructionEndpointUrl(id: string): string;
 
+    // FIXME: Deprecated
     abstract getName(): ApiName;
 
     abstract proceedDataToDishes(data: GenericDishStruct[], providedIngredients?: IngredientType[]): RatedDish[];
@@ -41,6 +44,8 @@ export abstract class AbstractApiService<GenericDishStruct, GenericIngredientStr
     abstract proceedDataToDishIngredients(ingredients: GenericIngredientStruct[]): DishIngredient[];
 
     abstract proceedDataToIngredientList(ingredients: GenericIngredientStruct[]): IngredientType[];
+
+    abstract getProvider(): DishProvider;
 
     async getDishes(ingredients: IngredientType[], mealType?: MealType): Promise<RatedDish[]> {
         const query = getQueryWithIngredientsAndDishType(ingredients, mealType, this.getName(), this.getApiKey());
@@ -76,7 +81,8 @@ export abstract class AbstractApiService<GenericDishStruct, GenericIngredientStr
         }
     }
 
-    async getDishDetails(id: string): Promise<DetailedDish> {
+    async getDishDetails(encodedDishId: EncodedDishId): Promise<DetailedDish> {
+        const { dishId: id } = DishIdObfuscator.decode(encodedDishId);
         const url: string = this.getFullApiUrl(this.getDishDetailsEndpointUrl(id));
         const context = 'AbstractApiService/getDishDetails';